Add basic API: continuity and smoothness.
Clean the version for DiffeomorphOn slightly; use symm when possible;
add a version of LocalDiffeomorphAt mimicking that.

Add a version for LocalDiffeomorph mimimicking that. I need to work a tiny
bit on one sorry, but that's basically re-using a proof I gave elsewhere.
Feels much nicer and cleaner now. \o/