refactor(Algebra/Algebra/Subalgebra/Basic): use a better defeq for `⊥ : Subalgebra R A` (#8038)
And the same thing for `StarSubalgebra R A`. `IntermediateField` was already handled in #7957.

As a result, nine (obvious) lemmas are now true by definition.

This slightly adjusts the statement of `Algebra.toSubmodule_bot` to make it simpler and true by definition; the original statement can be recovered by rewriting by `Submodule.one_eq_span`, which I've had to add in some downstream proofs.