chore(Order/BooleanAlgebra): reduce risk of instance diamonds (#8099)
We need `sup` and `max` to be defeq, and `max` is defined in Std. To achieve this, we define one in terms of the other.

Until #8105 was merged (which fixed up leanprover/std4#183), they were not defeq:
```lean
example :
  (GeneralizedCoheytingAlgebra.toLattice : Lattice Bool).sup =
    DistribLattice.toLattice.sup :=
  rfl -- failed
```

This instance in this PR should have been written such that it couldn't silently cause a diamond in the first place, by reusing existing data rather than creating new data.