Eliminate `erw` by using `change`
The real problem here is that the definition of `DirectSum.IsInternal.collectedBasis`
commits definitional abuse by using `DFinsupp.mapRange.linearEquiv` on terms which
are of type `DirectSum` rather than `DFinsupp`. One way to avoid this would be to
duplicate the API for `DFinsupp.mapRange` in terms of `DirectSum` and use this in
`collectedBasis`.