refactor(LinearAlgebra.ProjectiveSpace): localize notation and rename folder (#8394)
Following https://leanprover.zulipchat.com/#narrow/stream/116395-maths/topic/Projectivization I localize the notation $\mathbb{P}$ to the namespace `LinearAlgebra.Projectivization`. Following the discussion there, the namespace `Projectivization` is not optimal, as the notion can refer to different constructions, hence the choice of adding `LinearAlgebra.` in front of it.

Also, in accordance with the philosophy in the implementation, I replace "Projective Space" by "Projectivization" in the name of the folder, because the idea is that there is no "canonical" choice of the vector space one should start with to construct the projectivization.