chore: better phrasing of pointwise convergence of indicators (#8331)
In some recently added results, pointwise convergence of constant indicators of sets was phrased in terms of `Tendsto` in the product topology. The relevant notion is more elementary, however: it is (usually) equivalent to eventually constantness of the set membership of any point; in particular the topology of the underlying set plays almost no role (apart from mild separation axioms). This PR adds equivalences between the conditions and refactors some results.

Co-authored-by: Floris van Doorn  <@fpvandoorn>.