chore: merge master into nightly-testing via cron, not on every commit (#8627)
In the current setup, every commit to master is being merged into `nightly-testing`. This results in the cancellation of the previous CI job running on `nightly-testing`. If commits to master land often enough, this means that successful results never occur, and so the `nightly-testing-YYYY-MM-DD` branch is never created, even if `nightly-testing` would actually get green CI.

This changes to only merging master on a regular basis, long enough for CI to complete.



Co-authored-by: Scott Morrison <scott.morrison@gmail.com>