chore: build lean4checker with current toolchain (#8669)
After leanprover/lean4#2766 lands, the git hash of will be embedded in every olean. `lean4checker` will thus reject oleans unless it was compiled on the same toolchain as they were!

We may want to relax this later, but for now, since we are building `lean4checker` in CI anyway, let's just make sure it is on the same toolchain as mathlib.

Co-authored-by: Scott Morrison <scott.morrison@gmail.com>