feat: Mapping extreme points under continuous maps (#8574)
Prove that extreme points are preserved under affine equivalences, and the less trivial statement that a continuous affine map sends extreme points of a compact set to a superset of the extreme
points of the image of that set.

Also fix a few name and tweak the API a bit.




Co-authored-by: Yury G. Kudryashov <urkud@urkud.name>