Experiment with new more symmetric design.
The guiding principle of these changes was to make `RootPairing.flip_flip`
true by definition.

A failure of the previous design was that `RootPairing.flip_flip`
could not even be stated because the type of a `RootPairing` depended
on the type of the perfect pairing, and `LinearEquiv.flip_flip` is not
(and cannot be) true by definition.

There are two sets of changes here:
 1. Bundle the perfect pairing inside `RootPairing` so that its type
    no longer depends on the type of the perfect pairing.
 2. Introduce a new `PerfectPairing` type so that the `flip_flip` lemma
    can be true by definition.

Some tidy up remains to be done.