WIP: fix condition on IFT pregroupoids:
- to apply to the C^n pregroupoid, we should ask for some t' \sub t containing
x to have the pregroupoid property; we cannot expect this on all of t
(bijective need not imply invertible differential)
- we cannot generalise to a non-open neighbourhood, as the abstract theorem
wants a a local structo at x; the only witness we have comes from the IFT,
which yields a partial homeomorphisms with open source.

Abstract proof not fully fixed yet, but shouldn't be hard.