feat(Lindelöf Spaces): Add filter-definition
This adds the definition of Lindelöf spaces in terms of filters
and currently ends by proving that this definition agrees with the
more standard definition in terms of countable subcovers. Some basic
API is provided for working with Lindelof sets:
- Stability of standard operations, e.g. under certain intersections,
  images and subset
- IsLindelof.induction_on provides the important countable induction
  property for Lindelöf sets.
- IsLindelof.elim_countable_subcover provides existence of a countable
  subcover from a given open cover of a Lindelöf set. Some subsequent
  API applies this in various forms.
- isLindelof_iff_countable_subcover shows that the subcover definition
  is equivalent to the filter-definition used here.

Some easy extensions that have been omitted for the sake of length of
the PR and may be added at a later stage:
- Compact implies Lindelöf
- Second countable implies Lindelöf

Important omission: Compact.lean uses a few more intermediate results
in terms of directed families. Such an approach felt a bit unwieldy
to me for Lindelöf spaces, so I did not cover such intermediate results.

Point of consideration: the fact that we do not have CountableSet t
similar to how we have FinSet t makes some proofs slightly longer than
necessary.