New approach: use ContinuousLinearEquiv.ofBijective.
This is mathematically cleaner (I show injectivity and surjectivity of the
differential separately), cleaner in Lean (it's good practice to avoid heavy
tactic blocks in definitions) and makes all proofs by rfl work.

This requires completeness of the tangent spaces; this is slightly unfortunate,
as it was automatic in the previous approaches. Opinions welcome.

I'm a bit unsure about the local instances; we'll see if this can be avoided.