feat(Order/Filter/Basic) : Add Filter from prop
Add two definitions to create filter from a property on sets:
- Filter.ofProp_finite_inter if the property is stable under finite
  intersections and set enlargements.
- Filter.ofProp_finite_union if the property is stable under finite
  unions and set shrinking.

Filter.ofProp_finite_union is the key ingredient in
IsCompact.induction_on but may be convenient to have as individual
building block.

Both results also specialise to properties that are stable under
countable unions/intersections, in which case the filter is a
CountableInterFilter. This plays a role in IsLindelof.induction_on and
will be added when this PR is accepted.