fix: use `--install` flag in `lake update` hook (#9197)
Fixes an issue [reported on Zulip](https://leanprover.zulipchat.com/#narrow/stream/287929-mathlib4/topic/post-update.20hook.20failure/near/409543553). Without the `--install` flag elan will not download the new toolchain if it doesn't already have it. Since this is in an update call which is already expected to download things I think this should be okay.



Co-authored-by: Mario Carneiro <di.gama@gmail.com>