feat(CategoryTheory/Sites): objects which cover the terminal object (#8632)
In this PR, given a site `(C, J)`, we introduce the notion of a family of objects `Y : I → C` which "cover the final object". This notion is used in order to formulate results about global sections of sheaves. It shall also be useful in future PRs in order to formulate descent properties of sheaves.