feat(CategoryTheory/Shift): sequences of functors from a category with a shift (#9001)
Let `F : C ⥤ A` be a functor from a category `C` that is equipped with a shift by an additive monoid `M`. In this PR, we define a typeclass `F.ShiftSequence M` which includes the data of a sequence of functors `F.shift a : C ⥤ A` for all `a : A`. For each `a : A`, we have an isomorphism `F.isoShift a : shiftFunctor C a ⋙ F ≅ F.shift a` which satisfies some coherence relations. This will allow a better formulation of the long exact sequence attached to homological functors from a triangulated category to an abelian category.